Add bolt db as store option
